Add files via upload
In today's aviation industry, users or customers can book a seat or book a flight from anywhere in the world as long as they are connected to the Internet. Technology innovation has made air travel easier for customers who have a flight booking or reservation with just one click. So if thousands of people use planes on a daily basis to get from one place to another. Don't you think there should be a system to help this user book a flight, book a seat, pay for the flight, in their homes? . Therefore, we have designed an application that helps the customer to save his time and facilitate it to book a flight and displays the most seasonal tourist areas so that you can choose your destination easily. What distinguishes our application is that it is a comprehensive application for all the services that the user needs on his trip such as: the possibility of booking transportation, requesting a personal assistant, booking a tour guide, and. Reservation of accommodation in addition to booking a flight.